requirements.txt
amqp==5.3.1
asgiref==3.10.0
attrs==25.4.0
billiard==4.2.2
celery==5.5.3
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
Django==5.2.7
django-cors-headers==4.9.0 # 'corsheaders', 
django-jazzmin==3.0.1
django-redis==6.0.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dnspython==2.8.0
drf-spectacular==0.29.0
eventlet==0.40.3
greenlet==3.2.4
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kombu==5.5.4
packaging==25.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11 # psycopg2-binary
PyJWT==2.10.1
pytest==8.4.2
pytest-django==4.11.1
python-dateutil==2.9.0.post0
python-decouple==3.8
PyYAML==6.0.3
pillow==12.0.0
redis==7.0.1
referencing==0.37.0
rpds-py==0.28.0
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
vine==5.1.0
wcwidth==0.2.14
